Remove benchmark import guards
